Update aitext.json
Added more AI responses.

AI Detector is EXTREMELY accurate for 100% AI Generated responses, but is not so good with human responses. We hope by adding more AI data we will perfect it with what AI looks like, and it will be more confident with humans as if it doesn't know something with AI, then it would be human.